requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
baidu-aip==4.16.10
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
decorator==5.1.1
executing==1.2.0
Flask==2.2.3
idna==3.4
importlib-metadata==6.4.1
ipython==8.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
Naked==0.1.32
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pycrypto==2.6.1
Pygments==2.15.1
PyYAML==6.0
requests==2.28.2
shellescape==3.8.1
six==1.16.0
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.2.3
xmltodict==0.13.0
zipp==3.15.0
